Migrate the shell to shadcn dialogs, popovers and fields
Split dialogs, the bot panel and message cards out of Shell.tsx, replace the
hand-rolled create and user menus with popovers, fix the double-framed
choice rows, and remove the remaining hardcoded theme colours.
